requirements.txt
flake8==3.5.0
mccabe==0.6.1
mock==2.0.0
pbr==4.3.0
pycodestyle==2.3.1
pyflakes==1.6.0
six==1.11.0
python-docx
